requirements.txt
bleak==0.22.3
certifi==2025.10.5
charset-normalizer==3.4.3
dbus-fast==2.44.5
idna==3.10
meshcore==2.1.9
meshtastic==2.7.3
packaging==24.2
protobuf==6.32.1
pycayennelpp==2.4.0
Pypubsub==4.0.3
pyserial==3.5
pyserial-asyncio==0.6
PyYAML==6.0.3
requests==2.32.5
tabulate==0.9.0
urllib3==2.5.0
